CHANGE Help is now using Discord-pretty-help CHANGE Commands in cogs.
