+++ ark/requirements/common.txt
@@ -1,3 +1,4 @@
 discord==1.0.1
 requests==2.22.0
 python-decouple==3.1
+discord-pretty-help==1.2.1
