+++ ark/settings.py
@@ -1,7 +1,7 @@
 from decouple import config
 
 
-__version__ = '0.1.2'
+__version__ = '0.1.1'
 
 TOKEN = config('TOKEN', 'bot_token')
