+++ core/api_requests.py
@@ -1,7 +1,8 @@
 import requests
 from core.queries import (query_capsules, query_capsule, query_company,
-                          query_histories, query_history, query_missions,
-                          query_mission, query_rockets, query_rocket)
+                          query_histories, query_history, query_rockets, query_rocket)
+
+
 
 
 URL = 'https://api.spacex.land/graphql/'
@@ -50,36 +51,15 @@ def get_history(id):
     response = requests.post(URL, json={'query': query_history(id)})
 
     return response.json()
-
 def get_rockets():
     """
     request rockets information
     """
     response = requests.post(URL, json={'query': query_rockets()})
     return response.json()
-
 def get_rocket(id):
     """
     request single rocket information
     """
     response = requests.post(URL, json={'query': query_rocket(id)})
-    return response.json()
-
-
-def get_missions():
-    """
-    Requests missions data from API
-    """
-
-    response = requests.post(URL, json={'query': query_missions()})
-
-    return response.json()
-
-def get_mission(id):
-    """
-    Requests a single mission data from API by ID
-    """
-
-    response = requests.post(URL, json={'query': query_mission(id)})
-
     return response.json()
